+++ AnkhMorporkApp/Models/Fool.cs
@@ -4,9 +4,9 @@ namespace AnkhMorporkApp
     public class Fool 
     {
         public string Practice { get; set; }
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
 
-        public Fool(string practice, double fee)
+        public Fool(string practice, decimal fee)
         {
             this.Practice = practice;
             this.Fee = fee;
